Move ggrc_workflow's FE code into core module
