+++ src/ggrc/assets/javascripts/apps/workflow-config.js
