+++ src/ggrc/assets/javascripts/components/workflow/cycle-task-group-object-task/templates/cycle-task-group-object-task.mustache
