+++ src/ggrc/assets/javascripts/components/workflow/cycle-task-group-object-task/tests/cycle-task-group-object-task_spec.js
@@ -4,7 +4,7 @@
 */
 
 import Component from '../cycle-task-group-object-task';
-import RefreshQueue from '../../../../../../ggrc/assets/javascripts/models/refresh_queue';
+import RefreshQueue from '../../../../models/refresh_queue';
 
 describe('GGRC.Components.cycleTaskGroupObjectTask', function () {
   var viewModel;
