+++ src/ggrc/assets/javascripts/components/workflow/repeat-on-button.js
@@ -3,7 +3,7 @@
  Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
  */
 
-import '../../../../ggrc/assets/javascripts/components/dropdown/dropdown';
+import '../dropdown/dropdown';
 
 (function (can, GGRC, $) {
   'use strict';
