+++ src/ggrc/assets/javascripts/components/workflow/repeat-on-summary.js
