+++ src/ggrc/assets/javascripts/components/workflow/tests/workflow-start-cycle_spec.js
