+++ src/ggrc/assets/javascripts/components/workflow/workflow-start-cycle.js
