+++ src/ggrc/assets/javascripts/components/workflow/workflow_deactivate.js
