+++ src/ggrc/assets/javascripts/controllers/approval_workflow_controller.js
