+++ src/ggrc/assets/javascripts/controllers/dashboard_page.js
@@ -3,7 +3,7 @@
     Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
 */
 
-import RefreshQueue from '../../../../ggrc/assets/javascripts/models/refresh_queue';
+import RefreshQueue from '../models/refresh_queue';
 
 ;(function (CMS, GGRC, can, $) {
   GGRC.Components('dashboardWidgets', {
