+++ src/ggrc/assets/javascripts/controllers/tests/approval-workflow-model-helper_spec.js
