+++ src/ggrc/assets/javascripts/controllers/tests/dashbodrd_page_spec.js
