+++ src/ggrc/assets/javascripts/controllers/workflow_page.js
