+++ src/ggrc/assets/javascripts/entrypoints/modules.js
@@ -17,3 +17,21 @@ import '../apps/risk_assessments';
 
 // ggrc_risks
 import '../apps/risks';
+
+// ggrc_workflows
+import '../apps/workflow-config';
+import '../apps/workflows';
+import '../controllers/dashboard_page';
+import '../controllers/workflow_page';
+import '../workflow_mustache_helpers';
+import '../controllers/approval_workflow_controller';
+import '../components/workflow/workflow_clone';
+import '../components/workflow/taskgroup_clone';
+import '../components/workflow/end_cycle';
+import '../components/workflow/workflow-activate';
+import '../components/workflow/workflow-start-cycle';
+import '../components/workflow/workflow_deactivate';
+import '../components/workflow/repeat-on-button-wrapper';
+import '../components/workflow/repeat-on-button';
+import '../components/workflow/repeat-on-summary';
+import '../components/workflow/cycle-task-group-object-task/cycle-task-group-object-task';
