+++ src/ggrc/assets/javascripts/models/helpers.js
