+++ src/ggrc/assets/javascripts/models/index.js
@@ -53,3 +53,8 @@ import './risk_assessment';
 import './risk';
 import './risk_join_models';
 import './threat';
+import './cycle_models';
+import './task_group';
+import './workflow';
+import './workflow_join_models';
+import './helpers';
