+++ src/ggrc/assets/javascripts/models/tests/CycleTaskGroupObjectTask_model_spec.js
