+++ src/ggrc/assets/javascripts/models/workflow.js
