+++ src/ggrc/assets/javascripts/models/workflow_join_models.js
