+++ src/ggrc/assets/javascripts/plugins/utils/date-util.js
