--- src/ggrc_workflows/assets/javascripts/index.js
@@ -1,26 +1,0 @@
-/*!
-    Copyright (C) 2017 Google Inc.
-    Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
-*/
-
-import './apps/workflow-config';
-import './apps/workflows';
-import './controllers/dashboard_page';
-import './controllers/workflow_page';
-import './models/cycle_models';
-import './models/task_group';
-import './models/workflow';
-import './models/workflow_join_models';
-import './models/helpers';
-import './workflow_mustache_helpers';
-import './controllers/approval_workflow_controller';
-import './components/workflow_clone';
-import './components/taskgroup_clone';
-import './components/end_cycle';
-import './components/workflow-activate';
-import './components/workflow-start-cycle';
-import './components/workflow_deactivate';
-import './components/repeat-on-button-wrapper';
-import './components/repeat-on-button';
-import './components/repeat-on-summary';
-import './components/cycle-task-group-object-task/cycle-task-group-object-task';
